refactor: headerText typescript
