+++ src/components/BuildsModal.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { useEffect, useMemo } from 'react'
 import { Button, Card, Flex, Modal } from 'antd'
 import StatText from './characterPreview/StatText'
-import { HeaderText } from './HeaderText'
+import { HeaderText } from 'components/HeaderText'
 import { DeleteOutlined, ExclamationCircleOutlined } from '@ant-design/icons'
 import DB from '../lib/db'
 import { SaveState } from 'lib/saveState'
