+++ src/components/RelicFilterBar.jsx
@@ -2,7 +2,7 @@ import { Button, Flex, Select, theme, Tooltip, Typography } from 'antd'
 import React, { useEffect, useMemo, useState } from 'react'
 import { RelicScorer } from 'lib/relicScorerPotential'
 import CheckableTag from 'antd/lib/tag/CheckableTag'
-import { HeaderText } from './HeaderText'
+import { HeaderText } from 'components/HeaderText'
 import { TooltipImage } from './TooltipImage'
 import DB from '../lib/db'
 import { Hint } from 'lib/hint'
