+++ src/components/RelicModal.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components'
 import { Button, Flex, Form, Image, Input, InputNumber, Modal, Radio, Select, theme } from 'antd'
 import React, { ReactElement, useEffect, useMemo, useRef, useState } from 'react'
 import { Constants, setToId, Stats, UnreleasedSets } from 'lib/constants'
-import { HeaderText } from './HeaderText'
+import { HeaderText } from 'components/HeaderText'
 import { Message } from 'lib/message'
 import PropTypes from 'prop-types'
 import { Utils } from 'lib/utils'
