+++ src/components/optimizerTab/conditionals/CharacterConditionalDisplay.tsx
@@ -1,6 +1,6 @@
 import { Flex } from 'antd'
 import { memo } from 'react'
-import { HeaderText } from 'components/HeaderText.jsx'
+import { HeaderText } from 'components/HeaderText'
 import { TooltipImage } from 'components/TooltipImage.jsx'
 import { Hint } from 'lib/hint.jsx'
 import DisplayFormControl from 'components/optimizerTab/conditionals/DisplayFormControl'
@@ -40,12 +40,12 @@ export const CharacterConditionalDisplay = memo(({ id, eidolon, teammateIndex }:
   return (
     <Flex vertical gap={5}>
       {(teammateIndex == null)
-      && (
-        <Flex justify='space-between' align='center'>
-          <HeaderText>{t('CharacterPassives')/* Character passives */}</HeaderText>
-          <TooltipImage type={Hint.characterPassives()}/>
-        </Flex>
-      )}
+        && (
+          <Flex justify='space-between' align='center'>
+            <HeaderText>{t('CharacterPassives')/* Character passives */}</HeaderText>
+            <TooltipImage type={Hint.characterPassives()}/>
+          </Flex>
+        )}
       <DisplayFormControl content={content} teammateIndex={teammateIndex}/>
     </Flex>
   )
