+++ src/components/optimizerTab/optimizerForm/CharacterSelectorDisplay.tsx
@@ -1,5 +1,5 @@
 import { Flex, Form, Select } from 'antd'
-import { HeaderText } from 'components/HeaderText.jsx'
+import { HeaderText } from 'components/HeaderText'
 import { TooltipImage } from 'components/TooltipImage.jsx'
 import { Hint } from 'lib/hint.jsx'
 import RecommendedPresetsButton from 'components/optimizerTab/optimizerForm/RecommendedPresetsButton'
