+++ src/components/optimizerTab/optimizerForm/RelicMainSetFilters.tsx
@@ -1,5 +1,5 @@
 import { Button, Cascader, Flex, Form, Select } from 'antd'
-import { HeaderText } from 'components/HeaderText.jsx'
+import { HeaderText } from 'components/HeaderText'
 import { TooltipImage } from 'components/TooltipImage.jsx'
 import { Hint } from 'lib/hint.jsx'
 import { optimizerTabDefaultGap, panelWidth } from 'components/optimizerTab/optimizerTabConstants'
