+++ src/components/optimizerTab/optimizerForm/ResultFilters.tsx
@@ -1,5 +1,5 @@
 import { Flex } from 'antd'
-import { HeaderText } from 'components/HeaderText.jsx'
+import { HeaderText } from 'components/HeaderText'
 import { TooltipImage } from 'components/TooltipImage.jsx'
 import { Hint } from 'lib/hint.jsx'
 import FilterRow from 'components/optimizerTab/optimizerForm/FilterRow'
